[BACKEND] Towards a generic tcgen05.cp lowering (#8102)
This is the first PR towards a fully generic `tcgen05.cp` lowering. For
now it still has similar limitations as the previous lowering, but it
does not assume implicitly the layout of the shared memory and tensor
memory. Instead, it checks that the given TMEM and shmem layouts are
compatible with the instruction we are lowering to, and if so, computes
the matrix descriptor and tmem offsets manually.